fix update cart + add api update+delete (by id)

+++ cartservice/src/main/java/com/thc/cartservice/Repository/impl/CartRepositoryImpl.java
@@ -44,7 +44,7 @@ public class CartRepositoryImpl implements CartRepository {
     }
 
     @Override
-    public int count() {
+    public Long count() {
         return cartMapper.count();
     }
 }
