+++ cartservice/src/main/java/com/thc/cartservice/Service/CartService.java
@@ -18,5 +18,5 @@ public interface CartService {
     public int deleteCart(String id);
     public int deleteCartAll();
 
-    public int count();
+    public Long count();
 }
